Open window with previewUrl if it failed
In the desktop app `window.open( 'about:blank' )` fails.

Add a check for the preview window. If it doesn't exist when we have the previewUrl then directly open the window with the preview URL.

It may be possible to just create the window here in the first place but I am not sure if it is split for a reason