Adds a new action-only poller
This polling component can wrap another component in a polling mechanism
that runs a given function at a given interval and will optionally stop
polling if the component enters a hidden visibility state to save on
bandwidth.

If specified to pause when hidden, there is an optional parameter
describing whether or not the action should fire immediately upon coming
back into view. If not, it will fire after the given interval.

This component was inspired by the data-poller component but attempts to
be a more lightweight and readable component.