Fix poller to work
When I wrote the first draft of this component, I was doing so on a
plane without Internet access and I did not test it out. Not
surprisingly, it didn't work. This revision updates it to work the way
the original code was intended.

This does not change the functionality of the code, merely the syntax,
as the way I thought I might be able to apply the fat-arrow function
syntax didn't work in practice.